+++ data-manager/src/main/java/org/orbisgis/datamanager/postgis/PostgisTable.java
@@ -83,6 +83,25 @@ public class PostgisTable extends JdbcTable {
     }
 
     @Override
+    protected ResultSet getResultSet(){
+        if(resultSet == null) {
+            try {
+                resultSet = getStatement().executeQuery(getBaseQuery());
+            } catch (SQLException e) {
+                LOGGER.error("Unable to execute the query '"+getBaseQuery()+"'.\n"+e.getLocalizedMessage());
+                return null;
+            }
+            try {
+                resultSet.beforeFirst();
+            } catch (SQLException e) {
+                LOGGER.error("Unable to go before the first ResultSet row.\n" + e.getLocalizedMessage());
+                return null;
+            }
+        }
+        return resultSet;
+    }
+
+    @Override
     public Object asType(Class clazz) {
         if (clazz == ITable.class || clazz == PostgisTable.class) {
             return new PostgisTable(getTableLocation(), getBaseQuery(), (StatementWrapper)getStatement(),
